SEC-302 : Made account number generation mostly-truly random, also added luhn checksum as a bonus
